+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - added `myems_historical_db`.`tbl_text_value_latest` in database
 - added columns to `myems_fdd_db`.`tbl_web_messages` in database
 ### Changed
+- changed length of decimal part of latitude and longitude to 10 in database 
 - updated notification in myems-web
 - updated dockerfiles to reduce images size
 - updated command actions in myems-api